backend/python-services/drug-label-processor/requirements.txt
pymongo==4.12.1
groq
python-dotenv